Ultima versão que Igor pediu
sem tempo para escrever um bom commit

+++ firmware/esp/modules/traction_module/components/traction_comm/include/traction_comm.h
@@ -54,7 +54,7 @@ typedef struct {
     uint8_t pullup_enabled;
     uint32_t pwm_freq_hz;
     int32_t counts_per_motor_rev;
-    int32_t gear_ratio;
+    float gear_ratio;
     float rpm_max;
     char notes[64];
 } traction_comm_controller_cfg_state_t;
